Migrate to @tanstack/router
