+++ frontend/package.json
@@ -15,6 +15,7 @@
         "@eslint/js": "^9.35.0",
         "@microsoft/api-extractor": "^7.55.2",
         "@playwright/test": "^1.55.0",
+        "@tanstack/react-router-devtools": "^1.145.11",
         "@types/babel__standalone": "^7.1.9",
         "@types/d3": "^7.4.3",
         "@types/d3-sankey": "^0.12.4",
@@ -40,14 +41,13 @@
         "@mui/material": "^7.3.2",
         "@mui/x-data-grid": "^8.15.0",
         "@tanstack/react-query": "^5.66.0",
+        "@tanstack/react-router": "^1.144.0",
         "d3": "^7.9.0",
         "d3-sankey": "^0.12.3",
         "echarts": "^6.0.0",
         "react": "^19.1.1",
         "react-dom": "^19.1.1",
         "react-error-boundary": "^6.0.3",
-        "react-router": "^7.1.5",
-        "use-query-params": "^2.2.1",
         "zod": "^4.1.12"
     }
 }
