+++ frontend/src/components/Header.tsx
@@ -1,5 +1,5 @@
 import { Box } from "@mui/material";
-import { NavLink } from "react-router";
+import { Link } from "@tanstack/react-router";
 import { slugify } from "../router";
 import { useConfigContext } from "./ConfigProvider";
 
@@ -12,7 +12,9 @@ export function Header() {
     <Box className="headerline" sx={{ ".active": { color: "inherit" } }}>
       {dashboards.map((dashboard, i) => (
         <h3 key={i}>
-          <NavLink to={`/dashboards/${slugify(dashboard.name)}`}>{dashboard.name}</NavLink>
+          <Link to="." search={{ dashboard: slugify(dashboard.name) }}>
+            {dashboard.name}
+          </Link>
         </h3>
       ))}
     </Box>
